requirements.txt
Django==4.2.9
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-framework==0.1.0
django4-background-tasks==1.2.9
djangorestframework==3.14.0
earthengine-api==0.1.384
geemap==0.30.2
google-api-core==2.15.0
google-api-python-client==2.113.0
joblib==1.3.2
numpy==1.26.3
pandas==2.1.4
pipenv==2023.8.23
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
virtualenv==20.24.3

